Remove superseeded/old role transactions when a new role is granted to an address.

+++ src/script/bitcoinconsensus.cpp
@@ -92,7 +92,7 @@ static int verify_script(const unsigned char *scriptPubKey, unsigned int scriptP
             return set_error(err, bitcoinconsensus_ERR_TX_SIZE_MISMATCH);
 // FIXME
 std::cout << __func__ << ":" << __LINE__ << "> tx.nVersion=" << tx.nVersion;
-for (int i = 0; i < tx.vout.size(); ++i) {
+for (size_t i = 0; i < tx.vout.size(); ++i) {
     std::cout << " type=" << (int)(tx.vout[i].nTxType);
 }
 std::cout << std::endl;
